Adds generic annotation loader and updates Silex bridge.
This separates the annotation loading functionality (which is needed in any project) of the Silex bridge from the Silex-specific parts.

The new Schale\AnnotationLoader class can be registered as a loader for Doctrine Annotations' registry, but doesn't contain any Silex-specific functionality.

The remaining code in the Silex bridge's provider (AnnotationServiceProvider) uses this new AnnotationLoader class, and now just does the Silex-specific bits.